refactor variable name and package location as per CR nits
