+++ atlasdb-tests-shared/src/main/java/com/palantir/atlasdb/transaction/impl/TestTransactionManagerImpl.java
@@ -92,6 +92,6 @@ public class TestTransactionManagerImpl extends SerializableTransactionManager i
                 conflictDetectionManager.get(),
                 constraintModeSupplier.get(),
                 TransactionReadSentinelBehavior.THROW_EXCEPTION,
-                timestampCache);
+                timestampValidationReadCache);
     }
 }
